Basic working version in all platforms
